pyproject.toml
[tool.poetry]
name = "template-sign-up"
version = "0.1.0"
description = ""
authors = ["Akbota <aqbotamaratqyzy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular-sidecar = "^2024.4.1"
drf-spectacular = "^0.27.2"
djangorestframework = "^3.15.1"
openai = "^1.17.1"
environ = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
